make it an error when unknown inline HTML event handlers are found.
