+++ src/js/background.js
@@ -74,7 +74,6 @@ export function handleMessages(message, sender, sendResponse) {
       ORIGIN_ENDPOINT[message.origin] +
       '/' +
       message.version;
-    // TODO: Add error handling here
     fetch(endpoint, { METHOD: 'GET' })
       .then(response => response.json())
       .then(json => {
